made bike api call oe script. deleted db manager bc not needed with this set up.
