--- environment.yml
@@ -1,139 +1,0 @@
-name: software_eng
-channels:
-  - defaults
-dependencies:
-  - bzip2=1.0.8
-  - ca-certificates=2023.12.12
-  - libffi=3.4.4
-  - ncurses=6.4
-  - openssl=3.0.12
-  - pip=23.3.1
-  - python=3.11.7
-  - readline=8.2
-  - setuptools=68.2.2
-  - sqlite=3.41.2
-  - tk=8.6.12
-  - wheel=0.41.2
-  - xz=5.4.5
-  - zlib=1.2.13
-  - pip:
-      - anyio==4.2.0
-      - appnope==0.1.3
-      - argon2-cffi==23.1.0
-      - argon2-cffi-bindings==21.2.0
-      - arrow==1.3.0
-      - asttokens==2.4.1
-      - async-lru==2.0.4
-      - attrs==23.2.0
-      - babel==2.14.0
-      - beautifulsoup4==4.12.3
-      - bleach==6.1.0
-      - blinker==1.7.0
-      - cattrs==23.2.3
-      - certifi==2023.11.17
-      - cffi==1.16.0
-      - charset-normalizer==3.3.2
-      - click==8.1.7
-      - comm==0.2.1
-      - debugpy==1.8.0
-      - decorator==5.1.1
-      - defusedxml==0.7.1
-      - executing==2.0.1
-      - fastjsonschema==2.19.1
-      - flask==3.0.1
-      - flatbuffers==23.5.26
-      - fqdn==1.5.1
-      - greenlet==3.0.3
-      - idna==3.6
-      - ipykernel==6.29.0
-      - ipython==8.20.0
-      - ipywidgets==8.1.1
-      - isoduration==20.11.0
-      - itsdangerous==2.1.2
-      - jedi==0.19.1
-      - jinja2==3.1.3
-      - joblib==1.3.2
-      - json5==0.9.14
-      - jsonpointer==2.4
-      - jsonschema==4.21.1
-      - jsonschema-specifications==2023.12.1
-      - jupyter==1.0.0
-      - jupyter-client==8.6.0
-      - jupyter-console==6.6.3
-      - jupyter-core==5.7.1
-      - jupyter-events==0.9.0
-      - jupyter-lsp==2.2.2
-      - jupyter-server==2.12.5
-      - jupyter-server-terminals==0.5.2
-      - jupyterlab==4.0.11
-      - jupyterlab-pygments==0.3.0
-      - jupyterlab-server==2.25.2
-      - jupyterlab-widgets==3.0.9
-      - lxml==5.1.0
-      - markupsafe==2.1.4
-      - matplotlib-inline==0.1.6
-      - mistune==3.0.2
-      - nbclient==0.9.0
-      - nbconvert==7.14.2
-      - nbformat==5.9.2
-      - nest-asyncio==1.6.0
-      - notebook==7.0.7
-      - notebook-shim==0.2.3
-      - numpy==1.26.3
-      - openmeteo-requests==1.2.0
-      - openmeteo-sdk==1.10.0
-      - overrides==7.6.0
-      - packaging==23.2
-      - pandas==2.2.0
-      - pandocfilters==1.5.1
-      - parso==0.8.3
-      - pexpect==4.9.0
-      - platformdirs==4.1.0
-      - prometheus-client==0.19.0
-      - prompt-toolkit==3.0.43
-      - psutil==5.9.8
-      - ptyprocess==0.7.0
-      - pure-eval==0.2.2
-      - pycparser==2.21
-      - pygments==2.17.2
-      - python-dateutil==2.8.2
-      - python-dotenv==1.0.1
-      - python-json-logger==2.0.7
-      - pytz==2024.1
-      - pyyaml==6.0.1
-      - pyzmq==25.1.2
-      - qtconsole==5.5.1
-      - qtpy==2.4.1
-      - referencing==0.32.1
-      - requests==2.31.0
-      - requests-cache==1.1.1
-      - retry-requests==2.0.0
-      - rfc3339-validator==0.1.4
-      - rfc3986-validator==0.1.1
-      - rpds-py==0.17.1
-      - scikit-learn==1.4.0
-      - scipy==1.12.0
-      - send2trash==1.8.2
-      - six==1.16.0
-      - sniffio==1.3.0
-      - soupsieve==2.5
-      - sqlalchemy==2.0.26
-      - stack-data==0.6.3
-      - terminado==0.18.0
-      - threadpoolctl==3.2.0
-      - tinycss2==1.2.1
-      - tornado==6.4
-      - traitlets==5.14.1
-      - types-python-dateutil==2.8.19.20240106
-      - typing-extensions==4.9.0
-      - tzdata==2023.4
-      - uri-template==1.3.0
-      - url-normalize==1.4.3
-      - urllib3==2.1.0
-      - wcwidth==0.2.13
-      - webcolors==1.13
-      - webencodings==0.5.1
-      - websocket-client==1.7.0
-      - werkzeug==3.0.1
-      - widgetsnbextension==4.0.9
-prefix: /Users/adamsouthwick/miniconda3/envs/software_eng
